+++ scalingomock/domainsservice_mock.go
@@ -11,30 +11,30 @@ import (
 	gomock "github.com/golang/mock/gomock"
 )
 
-// MockDomainsService is a mock of DomainsService interface
+// MockDomainsService is a mock of DomainsService interface.
 type MockDomainsService struct {
 	ctrl     *gomock.Controller
 	recorder *MockDomainsServiceMockRecorder
 }
 
-// MockDomainsServiceMockRecorder is the mock recorder for MockDomainsService
+// MockDomainsServiceMockRecorder is the mock recorder for MockDomainsService.
 type MockDomainsServiceMockRecorder struct {
 	mock *MockDomainsService
 }
 
-// NewMockDomainsService creates a new mock instance
+// NewMockDomainsService creates a new mock instance.
 func NewMockDomainsService(ctrl *gomock.Controller) *MockDomainsService {
 	mock := &MockDomainsService{ctrl: ctrl}
 	mock.recorder = &MockDomainsServiceMockRecorder{mock}
 	return mock
 }
 
-// EXPECT returns an object that allows the caller to indicate expected use
+// EXPECT returns an object that allows the caller to indicate expected use.
 func (m *MockDomainsService) EXPECT() *MockDomainsServiceMockRecorder {
 	return m.recorder
 }
 
-// DomainSetCanonical mocks base method
+// DomainSetCanonical mocks base method.
 func (m *MockDomainsService) DomainSetCanonical(arg0, arg1 string) (scalingo.Domain, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "DomainSetCanonical", arg0, arg1)
@@ -43,13 +43,28 @@ func (m *MockDomainsService) DomainSetCanonical(arg0, arg1 string) (scalingo.Dom
 	return ret0, ret1
 }
 
-// DomainSetCanonical indicates an expected call of DomainSetCanonical
+// DomainSetCanonical indicates an expected call of DomainSetCanonical.
 func (mr *MockDomainsServiceMockRecorder) DomainSetCanonical(arg0, arg1 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DomainSetCanonical", reflect.TypeOf((*MockDomainsService)(nil).DomainSetCanonical), arg0, arg1)
 }
 
-// DomainUnsetCanonical mocks base method
+// DomainSetCertificate mocks base method.
+func (m *MockDomainsService) DomainSetCertificate(arg0, arg1, arg2, arg3 string) (scalingo.Domain, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "DomainSetCertificate", arg0, arg1, arg2, arg3)
+	ret0, _ := ret[0].(scalingo.Domain)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// DomainSetCertificate indicates an expected call of DomainSetCertificate.
+func (mr *MockDomainsServiceMockRecorder) DomainSetCertificate(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DomainSetCertificate", reflect.TypeOf((*MockDomainsService)(nil).DomainSetCertificate), arg0, arg1, arg2, arg3)
+}
+
+// DomainUnsetCanonical mocks base method.
 func (m *MockDomainsService) DomainUnsetCanonical(arg0 string) (scalingo.Domain, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "DomainUnsetCanonical", arg0)
@@ -58,13 +73,28 @@ func (m *MockDomainsService) DomainUnsetCanonical(arg0 string) (scalingo.Domain,
 	return ret0, ret1
 }
 
-// DomainUnsetCanonical indicates an expected call of DomainUnsetCanonical
+// DomainUnsetCanonical indicates an expected call of DomainUnsetCanonical.
 func (mr *MockDomainsServiceMockRecorder) DomainUnsetCanonical(arg0 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DomainUnsetCanonical", reflect.TypeOf((*MockDomainsService)(nil).DomainUnsetCanonical), arg0)
 }
 
-// DomainsAdd mocks base method
+// DomainUnsetCertificate mocks base method.
+func (m *MockDomainsService) DomainUnsetCertificate(arg0, arg1 string) (scalingo.Domain, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "DomainUnsetCertificate", arg0, arg1)
+	ret0, _ := ret[0].(scalingo.Domain)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// DomainUnsetCertificate indicates an expected call of DomainUnsetCertificate.
+func (mr *MockDomainsServiceMockRecorder) DomainUnsetCertificate(arg0, arg1 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DomainUnsetCertificate", reflect.TypeOf((*MockDomainsService)(nil).DomainUnsetCertificate), arg0, arg1)
+}
+
+// DomainsAdd mocks base method.
 func (m *MockDomainsService) DomainsAdd(arg0 string, arg1 scalingo.Domain) (scalingo.Domain, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "DomainsAdd", arg0, arg1)
@@ -73,13 +103,13 @@ func (m *MockDomainsService) DomainsAdd(arg0 string, arg1 scalingo.Domain) (scal
 	return ret0, ret1
 }
 
-// DomainsAdd indicates an expected call of DomainsAdd
+// DomainsAdd indicates an expected call of DomainsAdd.
 func (mr *MockDomainsServiceMockRecorder) DomainsAdd(arg0, arg1 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DomainsAdd", reflect.TypeOf((*MockDomainsService)(nil).DomainsAdd), arg0, arg1)
 }
 
-// DomainsList mocks base method
+// DomainsList mocks base method.
 func (m *MockDomainsService) DomainsList(arg0 string) ([]scalingo.Domain, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "DomainsList", arg0)
@@ -88,13 +118,13 @@ func (m *MockDomainsService) DomainsList(arg0 string) ([]scalingo.Domain, error)
 	return ret0, ret1
 }
 
-// DomainsList indicates an expected call of DomainsList
+// DomainsList indicates an expected call of DomainsList.
 func (mr *MockDomainsServiceMockRecorder) DomainsList(arg0 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DomainsList", reflect.TypeOf((*MockDomainsService)(nil).DomainsList), arg0)
 }
 
-// DomainsRemove mocks base method
+// DomainsRemove mocks base method.
 func (m *MockDomainsService) DomainsRemove(arg0, arg1 string) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "DomainsRemove", arg0, arg1)
@@ -102,13 +132,13 @@ func (m *MockDomainsService) DomainsRemove(arg0, arg1 string) error {
 	return ret0
 }
 
-// DomainsRemove indicates an expected call of DomainsRemove
+// DomainsRemove indicates an expected call of DomainsRemove.
 func (mr *MockDomainsServiceMockRecorder) DomainsRemove(arg0, arg1 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DomainsRemove", reflect.TypeOf((*MockDomainsService)(nil).DomainsRemove), arg0, arg1)
 }
 
-// DomainsUpdate mocks base method
+// DomainsUpdate mocks base method.
 func (m *MockDomainsService) DomainsUpdate(arg0, arg1, arg2, arg3 string) (scalingo.Domain, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "DomainsUpdate", arg0, arg1, arg2, arg3)
@@ -117,7 +147,7 @@ func (m *MockDomainsService) DomainsUpdate(arg0, arg1, arg2, arg3 string) (scali
 	return ret0, ret1
 }
 
-// DomainsUpdate indicates an expected call of DomainsUpdate
+// DomainsUpdate indicates an expected call of DomainsUpdate.
 func (mr *MockDomainsServiceMockRecorder) DomainsUpdate(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DomainsUpdate", reflect.TypeOf((*MockDomainsService)(nil).DomainsUpdate), arg0, arg1, arg2, arg3)